Complete generative suite

+++ pyrector/generative_providers.py
@@ -11,7 +11,8 @@ from PIL import Image
 
 import hashlib
 from abc import ABC
-from os import environ, makedirs
+from os import environ, makedirs, path
+from time import sleep
 
 
 class WithGoogleClient(ABC):
@@ -56,6 +57,10 @@ class WithElevenLabsClient(ABC):
 
 class GoogleTextProvider(WithGoogleClient, GenerativeTextProvider):
     def generate(self, prompt: str) -> str:
+        file_path = self.cache_file(prompt, f'{self.cache_dir}/text_', '.txt')
+        if path.exists(file_path):
+            return file_path
+
         if not self.client or not self.genai:
             raise ValueError('Google client not loaded')
 
@@ -65,11 +70,21 @@ class GoogleTextProvider(WithGoogleClient, GenerativeTextProvider):
         )
         if response.text is None:
             raise ValueError('No text returned from model')
-        return response.text.strip()
+
+        text = response.text.strip()
+
+        with open(file_path, 'w') as f:
+            f.write(text)
+
+        return text
 
 
 class GoogleImageProvider(WithGoogleClient, GenerativeImageProvider):
     def generate(self, prompt: str) -> str:
+        file_path = self.cache_file(prompt, f'{self.cache_dir}/image_', '.png')
+        if path.exists(file_path):
+            return file_path
+
         if not self.client or not self.genai:
             raise ValueError('Google client not loaded')
 
@@ -86,9 +101,6 @@ class GoogleImageProvider(WithGoogleClient, GenerativeImageProvider):
         if response.parts is None:
             raise ValueError('No content returned from model')
 
-        prompt_hash = hashlib.sha256(prompt.encode('utf-8')).hexdigest()
-        file_path = f'{self.cache_dir}/image_{prompt_hash}.png'
-
         for part in response.parts:
             if part.inline_data is not None:
                 image = part.as_image()
@@ -100,10 +112,71 @@ class GoogleImageProvider(WithGoogleClient, GenerativeImageProvider):
         raise ValueError('No image returned from model')
 
 
+class GoogleVideoProvider(WithGoogleClient, GenerativeVideoProvider):
+    def generate(self, prompt: str) -> str:
+        file_path = self.cache_file(prompt, f'{self.cache_dir}/video_', '.mp4')
+        if path.exists(file_path):
+            return file_path
+
+        if not self.client or not self.genai:
+            raise ValueError('Google client not loaded')
+
+        operation = self.client.models.generate_videos(
+            model="veo-3.1-fast-generate-preview",
+            prompt=prompt,
+            config=self.genai.types.GenerateVideosConfig(
+                aspect_ratio='16:9',  # TODO: derive aspect ratio
+                duration_seconds=8,
+                resolution='720p',
+            ),
+        )
+
+        while not operation.done:
+            print("Waiting for video generation to complete...")
+            sleep(5)
+            operation = self.client.operations.get(operation)
+
+        if (not operation.response or not operation.response.generated_videos):
+            raise ValueError('No video returned from model')
+
+        generated_video = operation.response.generated_videos[0]
+        if not generated_video.video:
+            raise ValueError('No video returned from model')
+
+        self.client.files.download(file=generated_video.video)
+        generated_video.video.save(file_path)
+
+        return file_path
+
+
+class ElevenLabsAudioProvider(WithElevenLabsClient, GenerativeAudioProvider):
+    def generate(self, prompt: str) -> str:
+        file_path = self.cache_file(prompt, f'{self.cache_dir}/audio_', '.mp3')
+        if path.exists(file_path):
+            return file_path
+
+        if not self.client or not self.elevenlabs_client:
+            raise ValueError('ElevenLabs client not loaded')
+
+        response = self.client.text_to_sound_effects.convert(
+            text=prompt,
+        )
+
+        response_bytes = b''.join(response)
+        with open(file_path, 'wb') as f:
+            f.write(response_bytes)
+
+        return file_path
+
+
 class ElevenLabsVoiceProvider(WithElevenLabsClient, GenerativeVoiceProvider):
     def generate(self, prompt: str) -> str:
         if not self.client or not self.elevenlabs_client:
             raise ValueError('ElevenLabs client not loaded')
+
+        file_path = self.cache_file(prompt, f'{self.cache_dir}/voice_', '.mp3')
+        if path.exists(file_path):
+            return file_path
 
         response = self.client.text_to_speech.convert(
             voice_id='4nLP0u2B3yI0lyzATFnN',  # TODO: make this configurable
@@ -111,33 +184,30 @@ class ElevenLabsVoiceProvider(WithElevenLabsClient, GenerativeVoiceProvider):
             model_id='eleven_turbo_v2_5',
         )
 
-        prompt_hash = hashlib.sha256(prompt.encode('utf-8')).hexdigest()
-        file_path = f'{self.cache_dir}/speech_{prompt_hash}.mp3'
-
+        response_bytes = b''.join(response)
         with open(file_path, 'wb') as f:
-            for chunk in response:
-                if chunk:
-                    f.write(chunk)
+            f.write(response_bytes)
 
         return file_path
 
 
 class ElevenLabsMusicProvider(WithElevenLabsClient, GenerativeMusicProvider):
     def generate(self, prompt: str) -> str:
+        file_path = self.cache_file(prompt, f'{self.cache_dir}/music_', '.mp3')
+        if path.exists(file_path):
+            return file_path
+
         if not self.client or not self.elevenlabs_client:
             raise ValueError('ElevenLabs client not loaded')
 
         response = self.client.music.compose(
             prompt=prompt,
             model_id='music_v1',
+            music_length_ms=30_000,
         )
 
-        prompt_hash = hashlib.sha256(prompt.encode('utf-8')).hexdigest()
-        file_path = f'{self.cache_dir}/music_{prompt_hash}.mp3'
-
+        response_bytes = b''.join(response)
         with open(file_path, 'wb') as f:
-            for chunk in response:
-                if chunk:
-                    f.write(chunk)
+            f.write(response_bytes)
 
         return file_path
